[TECH-521] Move build logic into its own command package
branch: feature/TECH-521-simplified-remote-artifact-caching
